requirements.txt
alpaca-trade-api
yfinance
pandas
numpy
scikit-learn
ta
python-dotenv
requests
beautifulsoup4
nltk
newsapi-python
matplotlib
streamlit
textblob
plotly